feat: extract the useObserverTable hook
